requirements.txt
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
Django==4.1.7
django-multiselectfield==0.1.12
djangorestframework==3.14.0
driver==0.1.0
dynaconf==3.2.0
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.7.1
fonttools==4.38.0
fqdn==1.5.1
h11==0.14.0
html5lib==1.1
hvac==1.1.1
idna==3.4
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
knox==0.1.14
kombu==5.3.0
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.5
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
outcome==1.2.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyhcl==0.4.4
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
redis==4.5.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.10.0
Send2Trash==1.8.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2022.7
uri-template==1.2.0
urllib3==2.0.2
validators==0.20.0
vine==5.0.0
virtualenv==20.16.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
win32-setctime==1.1.0
wsproto==1.2.0
